♻️ refactor modal components
