+++ src/commands/announcements/add.ts
@@ -1,10 +1,10 @@
-import { EmbedLimits, TextInputLimits } from '@sapphire/discord-utilities'
 import { fetchT, TFunction } from '@sapphire/plugin-i18next'
 import { Subcommand } from '@sapphire/plugin-subcommands'
 import { MessageAttachment, Modal } from 'discord.js'
 import ow from 'ow'
-import { MessageComponentTypes, TextInputStyles } from 'discord.js/typings/enums'
+import { MessageComponentTypes } from 'discord.js/typings/enums'
 import { nameSchema } from '../../schemas/OwSchemas'
+import getModalComponents from '../../utils/getModalComponents'
 import { Image, temporaryImgStorage } from '../../utils/Globals'
 import { validateChatInput } from '../../utils/validateOptions'
 
@@ -40,58 +40,7 @@ export async function add (interaction: Subcommand.ChatInputInteraction) {
     .setTitle(t('commands:add.modalTitle'))
     .setCustomId(`addAnnouncement:${interaction.id}:${Date.now()}:${id}`)
 
-  const components = [
-    /*
-      {
-        customId: 'name',
-        label: 'Name',
-        maxLength: 16,
-        minLength: 3,
-        placeholder: 'Escriba el nombre del anuncio, este valor es puramente identificativo',
-        required: true,
-        style: TextInputStyles.SHORT,
-        type: MessageComponentTypes.TEXT_INPUT
-      }
-    */
-    {
-      customId: 'title',
-      maxLength: EmbedLimits.MaximumTitleLength,
-      required: false,
-      style: TextInputStyles.PARAGRAPH,
-      type: MessageComponentTypes.TEXT_INPUT
-    },
-    {
-      customId: 'description',
-      maxLength: TextInputLimits.MaximumValueCharacters,
-      required: true,
-      style: TextInputStyles.PARAGRAPH,
-      type: MessageComponentTypes.TEXT_INPUT
-    },
-    {
-      customId: 'footer',
-      maxLength: EmbedLimits.MaximumFooterLength,
-      required: false,
-      style: TextInputStyles.PARAGRAPH,
-      type: MessageComponentTypes.TEXT_INPUT
-    },
-    {
-      customId: 'url',
-      required: false,
-      style: TextInputStyles.SHORT,
-      type: MessageComponentTypes.TEXT_INPUT
-    },
-    {
-      customId: 'color',
-      required: false,
-      style: TextInputStyles.SHORT,
-      type: MessageComponentTypes.TEXT_INPUT
-    }
-  ]
-    .map((component) => ({
-      ...component,
-      label: t(`commands:add.modal.${component.customId}.label`),
-      placeholder: t(`commands:add.modal.${component.customId}.placeholder`)
-    }))
+  const components = await getModalComponents(interaction)
 
   // @ts-expect-error
   modal.setComponents([
